+++ sdk/quota/arm-quota/src/azureQuotaExtensionAPI.ts
@@ -6,6 +6,7 @@
  * Changes may cause incorrect behavior and will be lost if the code is regenerated.
  */
 
+import * as coreClient from "@azure/core-client";
 import * as coreAuth from "@azure/core-auth";
 import {
   UsagesImpl,
@@ -19,10 +20,12 @@ import {
   QuotaRequestStatus,
   QuotaOperation
 } from "./operationsInterfaces";
-import { AzureQuotaExtensionAPIContext } from "./azureQuotaExtensionAPIContext";
 import { AzureQuotaExtensionAPIOptionalParams } from "./models";
 
-export class AzureQuotaExtensionAPI extends AzureQuotaExtensionAPIContext {
+export class AzureQuotaExtensionAPI extends coreClient.ServiceClient {
+  $host: string;
+  apiVersion: string;
+
   /**
    * Initializes a new instance of the AzureQuotaExtensionAPI class.
    * @param credentials Subscription credentials which uniquely identify client subscription.
@@ -32,7 +35,41 @@ export class AzureQuotaExtensionAPI extends AzureQuotaExtensionAPIContext {
     credentials: coreAuth.TokenCredential,
     options?: AzureQuotaExtensionAPIOptionalParams
   ) {
-    super(credentials, options);
+    if (credentials === undefined) {
+      throw new Error("'credentials' cannot be null");
+    }
+
+    // Initializing default values for options
+    if (!options) {
+      options = {};
+    }
+    const defaults: AzureQuotaExtensionAPIOptionalParams = {
+      requestContentType: "application/json; charset=utf-8",
+      credential: credentials
+    };
+
+    const packageDetails = `azsdk-js-arm-quota/1.0.0-beta.1`;
+    const userAgentPrefix =
+      options.userAgentOptions && options.userAgentOptions.userAgentPrefix
+        ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
+        : `${packageDetails}`;
+
+    if (!options.credentialScopes) {
+      options.credentialScopes = ["https://management.azure.com/.default"];
+    }
+    const optionsWithDefaults = {
+      ...defaults,
+      ...options,
+      userAgentOptions: {
+        userAgentPrefix
+      },
+      baseUri: options.endpoint || "https://management.azure.com"
+    };
+    super(optionsWithDefaults);
+
+    // Assigning values to Constant parameters
+    this.$host = options.$host || "https://management.azure.com";
+    this.apiVersion = options.apiVersion || "2021-03-15-preview";
     this.usages = new UsagesImpl(this);
     this.quota = new QuotaImpl(this);
     this.quotaRequestStatus = new QuotaRequestStatusImpl(this);